refactor: changed file structure
