+++ Ipsen5_groep01_frontend/Components/Layout/NavMenu.razor.css
@@ -1,24 +1,23 @@
 .navbar {
     width: 100%;
-    min-height: 100%;
-    height: auto;
+    
+
     display: flex;
     align-items: center;
-    padding: 0;
+   
 }
 
 .staffingnow_logo {
     width: 100%;
-    max-width: 330px;
+    width: 23vmin;
     margin-left: 5%;
 }
 
 .profile_pic {
-    height: 130px;
-    max-width: 100%;
+    height: 8vmin;
 }
 
-.admin-buttons, .home-buttons {
+.admin-buttons {
     flex: 1;
     height: 100%;
     display: flex;
@@ -28,18 +27,29 @@
     gap: 3%;
 }
 
-.home-buttons{
-   justify-content: right;
-   margin-right: 15%;
+
+.home-buttons {
+    display: flex;
+    justify-content: end;
+    margin-right: 10vmin;
+    gap: 3vmin
 }
 
 
-.button {
+.signInButton {
     display: flex;
     justify-content: center;
     align-items: center;
-    min-width: 200px;
-    height: 50px;
+    width: 18vmin;
+    height: 4vmin;
+}
+
+.registerButton {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 18vmin;
+    height: 4vmin;
 }
