+++ Ipsen5_groep01_frontend/Components/Pages/AccountBeheer.razor
@@ -1,5 +1,6 @@
 @page "/AccountBeheer"
-@using Services
+@using Services;
+@using Ipsen5_groep01_frontend.Services
 @inject LoginService LoginService
 
 <h1>AccountBeheer</h1>
