+++ Ipsen5_groep01_frontend/Components/Pages/DocumentBeheer.razor
@@ -1,8 +1,8 @@
 ﻿@page "/documentbeheer"
 @attribute [StreamRendering]
-@using Ipsen5_groep01_frontend.Components.Services
+@using Ipsen5_groep01_frontend.Services
 @using Ipsen5_groep01_frontend.Models
-@inject Ipsen5_groep01_frontend.Components.Services.UploadTypeService UploadTypeService
+@inject UploadTypeService UploadTypeService
 @inject LoginService LoginService
 
 <div class="decorative-bar" style="width:100%; height: 70px; background: linear-gradient(0.25turn, #FF6700 0%, #1A222D 53%, #183749 70%, #10ADE5 100%); margin-bottom: 30px;"></div>
