+++ Ipsen5_groep01_frontend/Components/Pages/DossierAanmaken.razor
@@ -3,7 +3,8 @@
 @inject NavigationManager NavigationManager
 @using Ipsen5_groep01_frontend.Components.Data
 
-@using Services
+@using Services;
+@using Ipsen5_groep01_frontend.Services
 @inject LoginService LoginService
 
 <PageTitle>Dossier Aanmaken</PageTitle>
