+++ Ipsen5_groep01_frontend/Components/Pages/Home.razor
@@ -1,7 +1,8 @@
 ﻿@page "/"
 @inject NavigationManager NavigationManager
 
-@using Services
+@using Services;
+@using Ipsen5_groep01_frontend.Services
 @inject LoginService LoginService
