+++ Ipsen5_groep01_frontend/Components/Pages/Home.razor.css
@@ -16,10 +16,6 @@
     padding-top: 2.3vmin;
 }
 
-.logo {
-    height: 6vmin
-}
-
 .sign-container {
     display: flex;
     justify-content: end;
@@ -71,7 +67,7 @@
     display: flex;
     justify-content: center; 
     align-items: center;
-    height: 600px;
+    height: 65vmin;
 }
 
 
@@ -108,7 +104,7 @@
 .footer-text-container {
     display: flex;
     align-items: center;
-    justify-content: center;
+    padding-left: 5vmin;
 }
 
 .footer-text {
