+++ Ipsen5_groep01_frontend/Components/Pages/InterneMedewerkerDossier.razor.css
@@ -128,3 +128,44 @@
 .table-row.green-background:hover td {
     background-color: #4eb42b; /* Adjust this color as needed */
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
