+++ Ipsen5_groep01_frontend/Components/Pages/LeverancierDossierInzien.razor
@@ -4,7 +4,8 @@
 @using Models
 @inject NavigationManager NavigationManager
 
-@using Services
+@using Services;
+@using Ipsen5_groep01_frontend.Services
 @inject LoginService LoginService
 @inject FileUploadService FileUploadService
 @inject IJSRuntime JSRuntime
