+++ Ipsen5_groep01_frontend/Components/Pages/NewDocument.razor
@@ -1,7 +1,7 @@
 @page "/NewDocument"
 @attribute [StreamRendering]
-@using Ipsen5_groep01_frontend.Components.Services
-@inject Ipsen5_groep01_frontend.Components.Services.UploadTypeService UploadTypeService
+@using Ipsen5_groep01_frontend.Services
+@inject UploadTypeService UploadTypeService
 @inject NavigationManager NavigationManager
 
 @using Ipsen5_groep01_frontend.Models
