+++ Ipsen5_groep01_frontend/Components/Pages/Professionals.razor
@@ -1,5 +1,6 @@
 @page "/Professionals"
-@using Services
+@using Services;
+@using Ipsen5_groep01_frontend.Services
 @inject LoginService LoginService
 
 <h1>Professionals</h1>
