+++ Ipsen5_groep01_frontend/Components/Pages/SignIn.razor
@@ -11,27 +11,32 @@
 @using Microsoft.AspNetCore.Identity;
 @using Microsoft.IdentityModel.Tokens;
 @using System.IdentityModel.Tokens.Jwt;
+@using Ipsen5_groep01_frontend.Services;
 @using System.Net;
 @using System.Runtime.InteropServices;
 @using System.Security.Claims;
 @using System.Text;
 @using static System.Runtime.InteropServices.JavaScript.JSType;
-@using System.ComponentModel.DataAnnotations
-@using Microsoft.AspNetCore.Components.Forms
-@using Ipsen5_groep01_frontend.Models
-@using Ipsen5_groep01_frontend.Services
+@using System.ComponentModel.DataAnnotations;
+@using Microsoft.AspNetCore.Components.Forms;
+@using Ipsen5_groep01_frontend.Models;
+@using Ipsen5_groep01_frontend.Services;
+
 @rendermode InteractiveAuto
 
 <PageTitle>Inloggen</PageTitle>
-
-
 
 <div class="page">
     <div class="linear-gradient-Container">
 
         <div class="login-container">
             <div class="title-login">
-                INLOGGEN
+                <div class="title-login">
+                    <div class="inlog-text">Inlog</div>
+                    <div class="gen-text">gen</div>
+                </div>
+
+
             </div>
             <div class="login">
 
@@ -45,9 +50,6 @@
                             <InputText id="email" @bind-Value="SignInModel.Email" class="form-control"></InputText>
                             <ValidationMessage For="() => SignInModel.Email"></ValidationMessage>
                         </div>
-
-
-
 
                         <div class="form-group">
                             <label>Password:</label>
