+++ Ipsen5_groep01_frontend/Ipsen5_groep01_frontend.csproj
@@ -7,9 +7,11 @@
         <DockerDefaultTargetOS>Windows</DockerDefaultTargetOS>
     </PropertyGroup>
 
-
     <ItemGroup>
-      <Folder Include="Components\NewFolder\" />
+      <Compile Remove="Components\NewFolder\**" />
+      <Content Remove="Components\NewFolder\**" />
+      <EmbeddedResource Remove="Components\NewFolder\**" />
+      <None Remove="Components\NewFolder\**" />
     </ItemGroup>
 
     <ItemGroup>
@@ -24,6 +26,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <Folder Include="Components\Services\" />
       <Folder Include="wwwroot\images\" />
     </ItemGroup>
