+++ Ipsen5_groep01_frontend/Models/Dossier/DuurOvereenkomst.cs
