+++ Ipsen5_groep01_frontend/Services/LoginService.cs
@@ -1,27 +1,31 @@
 using System.Threading.Tasks;
 using System.Collections.Generic;
 
-namespace Ipsen5_groep01_frontend.Components.Services
+namespace Ipsen5_groep01_frontend.Services
 {
     public class LoginService
     {
         private bool loggedIn = false;
         private bool isAdmin = false;
 
-        public bool getLoggedIn(){
+        public bool getLoggedIn()
+        {
             return loggedIn;
         }
 
-        public bool getIsAdmin(){
+        public bool getIsAdmin()
+        {
             return isAdmin;
         }
 
-        public void setLoggedIn(bool value){
-            this.loggedIn = value;
+        public void setLoggedIn(bool value)
+        {
+            loggedIn = value;
         }
 
-        public void setIsAdmin(bool value){
-            this.isAdmin = value;
+        public void setIsAdmin(bool value)
+        {
+            isAdmin = value;
         }
 
         public void ResetState()
